Adding Zod for the edit and checkout. Also create a Zod schema for the pet id

+++ src/actions/actions.ts
@@ -6,9 +6,9 @@ import prisma from "@/lib/db";
 import { sleep } from "@/lib/utils";
 import { PetEssentials } from "@/lib/types";
 import { Pet } from "@prisma/client";
-import { petFormSchema } from "@/lib/validations";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
 
   const validatedPet = petFormSchema.safeParse(pet);
@@ -31,12 +31,21 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], petData: PetEssentials) {
+export async function editPet(petId: unknown, petData: unknown) {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(petData);
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPet.success || !validatedPetId.success) {
+    return {
+      message: "Invalid pet data."
+    };
+  }
+
   try {
     await prisma.pet.update({
-      where: { id: petId },
-      data: petData
+      where: { id: validatedPetId.data },
+      data: validatedPet.data
     });
   } catch (error) {
     return {
@@ -46,11 +55,19 @@ export async function editPet(petId: Pet["id"], petData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: Pet["id"]) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data."
+    };
+  }
+
   try {
     await prisma.pet.delete({
-      where: { id: petId }
+      where: { id: validatedPetId.data }
     });
   } catch (error) {
     return {
